chore: moved types/sign to shared
