--- src/@types/flatley/index.d.ts
@@ -1,19 +1,0 @@
-/** Declaration file generated by dts-gen */
-
-declare module "flatley" {
-  function flatten(target: any, opts: FlatleyOptions): any;
-
-  interface Coercion {
-    test: (key: string, value: any) => boolean;
-    transform?: (value: any) => any;
-  }
-
-  interface FlatleyOptions {
-    delimiter?: string;
-    safe?: boolean;
-    object?: boolean;
-    overwrite?: boolean;
-    maxDepth?: number;
-    coercion?: Coercion[];
-  }
-}
