+++ src/shared/privacy/privacy.test.ts
@@ -1,5 +1,5 @@
 import { obfuscateData, obfuscateDocument } from "./privacy";
-import { WrappedDocument, SchemaId } from "../../@types/document";
+import { WrappedDocument, SchemaId } from "../../shared/@types/document";
 import { getData } from "../utils";
 
 describe("privacy", () => {
