+++ src/shared/sign/ecdsa-secp256k1-signature-2019.ts
