+++ src/shared/utils/utils.test.ts
@@ -1,6 +1,6 @@
 import * as utils from "./utils";
-import { VerifiableCredential, wrapDocument } from "../../";
-import { SchemaId, WrappedDocument } from "../../@types/document";
+import { wrapDocument } from "../../";
+import { VerifiableCredential, SchemaId, WrappedDocument } from "../../shared/@types/document";
 import { IdentityProofType, OpenAttestationDocument as v2OpenAttestationDocument } from "../../__generated__/schemaV2";
 import { IdentityType, Method, OpenAttestationDocument, ProofType, TemplateType } from "../../__generated__/schemaV3";
