+++ src/v2/signature/signature.ts
@@ -3,7 +3,7 @@ import { keccak256 } from "js-sha3";
 import { digestDocument } from "../../shared/digest";
 import { MerkleTree } from "../../shared/merkle";
 import { hashToBuffer, bufSortJoin } from "../../shared/utils";
-import { SchematisedDocument, Signature, WrappedDocument } from "../../@types/document";
+import { SchematisedDocument, Signature, WrappedDocument } from "../../shared/@types/document";
 import { OpenAttestationDocument } from "../../__generated__/schemaV3";
 
 export const wrap = <T = OpenAttestationDocument>(
