+++ src/v3/signature/signature.ts
@@ -2,7 +2,7 @@ import { OpenAttestationDocument } from "../../__generated__/schemaV3";
 import { bufSortJoin, hashToBuffer, toBuffer } from "../../shared/utils";
 import { v4 as uuid } from "uuid";
 import { MerkleTree } from "../../shared/merkle";
-import { Salt, VerifiableCredential } from "../../@types/document";
+import { Salt, VerifiableCredential } from "../../shared/@types/document";
 import { cloneDeep, get, sortBy, unset } from "lodash";
 import { keccak256 } from "js-sha3";
 import { compact } from "jsonld";
